// src/types/certificate.ts
export type CertificateData = {
  recipientName: string;
  courseTitle: string;
  academyName: string;
  verificationHash: string;
  certificateId: string;
  auditorTitle: string;
  auditorOrg: string;
  mentorTitle: string;
  mentorOrg: string;
  auditorSignatureUrl: string;
  mentorSignatureUrl: string;
};